==> 90.UnnaturalPTM/01.Cys_Vs_CMC.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r echo=FALSE}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(gridExtra)
library(ggthemes)
library(ggpmisc)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(ggExtra)
library(dplyr)
```

```{r echo=FALSE}
get_density <- function(x, y, n = 1000, ...) {
  dens <- MASS::kde2d(x, y, n = n, ...)
  ix <- findInterval(x, dens$x)
  iy <- findInterval(y, dens$y)
  ii <- cbind(ix, iy)
  return(dens$z[ii] / max(dens$z[ii]))
}
```


```{r}
sig1 <- fread(paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals/", "Cys", ".signal.txt"))
sig2 <- fread(paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals/", "CbC", ".signal.txt"))
sig12 <- rbind(sig1, sig2)
sig12[, AA := gsub("CbC", "CMC", AA)]
sig12[, AA := factor(AA, levels = c("Cys", "CMC"))]
```

```{r}
ggplot(sig12[State == "Sington"], aes(x = Blockade, y = DwellTime, colour = AA)) + 
  geom_point(aes(shape = as.character(File)), alpha = 0.5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0, 1)) + 
  guides(colour = guide_legend(nrow = 2, reverse = F), shape = "none") + 
  labs(y = "Dwell time (ms)") + 
  theme_light(base_size = 15) + 
  theme(plot.tag = element_blank(), 
        legend.title = element_blank(), 
        legend.background = element_blank(), 
        legend.box.spacing = unit(0, "line"), 
        panel.grid = element_blank(), 
        legend.box.background = element_blank(), 
        plot.margin = unit(c(0, 1, 1, 1), "line"),
        legend.position = "right") -> p1
p1
```

```{r}
ggplot(sig12[State != "Sington"], aes(x = Blockade, y = DwellTime, colour = AA)) + 
  geom_point(aes(shape = as.character(File)), alpha = 0.5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0, 1)) + 
  guides(colour = guide_legend(nrow = 2, reverse = F), shape = "none") + 
  labs(y = "Dwell time (ms)") + 
  theme_light(base_size = 15) + 
  theme(plot.tag = element_blank(), 
        legend.title = element_blank(), 
        legend.background = element_blank(), 
        legend.box.spacing = unit(0, "line"), 
        panel.grid = element_blank(), 
        legend.box.background = element_blank(), 
        plot.margin = unit(c(0, 1, 1, 1), "line"),
        legend.position = "right") -> p2
p2
```

```{r}
abf_Cys <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_21504011.Rds"))
abf_CMC <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_20230404_0005.Rds"))
```

```{r}
sig12[State == "Sington" & File == "21504011"]
```

```{r}
i <- sig12[State == "Sington" & File == "21504011"][5, ID]
ggplot(abf_Cys[Time > sig12[ID == i, StartTime - 0.005] & Time < sig12[ID == i, EndTime + 0.005]], aes(x = Time, y = pA)) + 
  geom_step() + 
  theme_base()
```

```{r}
i <- sig12[State == "Sington" & File == "20230404_0005"][3, ID]
ggplot(abf_CMC[Time > sig12[ID == i, StartTime - 0.005] & Time < sig12[ID == i, EndTime + 0.005]], aes(x = Time, y = pA)) + 
  geom_step() + 
  theme_base()
```




```{r}
files <- list.files("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals", full.names = T)
sig0 <- mclapply(files, fread)
sig0 <- data.table(A = rep(gsub(".signal.txt", "", basename(files)), mapply(nrow, sig0)), do.call(rbind, sig0))
```

```{r}
set.seed(12345)
sigi <- rbind(sig0[State == "Sington" & DwellTime > 1 & SignalCurrentPercent > 80, .SD[sample(.N, 1)], .(A, State)], 
              sig0[State == "Mixtrue" & DwellTime > 1, .SD[sample(.N, 1)], .(A, State)])
```

```{r}
sigi_abf <- mclapply(seq_len(nrow(sigi)), function(i) {
  abf <- readRDS(sigi[i, paste0("./analysis/81.ABFProcessing/ABF/ABF_", File, ".Rds")])
  abf <- abf[Time > sigi[i, StartTime] - 0.001 & Time < sigi[i, EndTime] + 0.001]
  data.table(sigi[i, .(A, AA, State, ID)], abf)
}, mc.cores = 11)
sigi_abf <- do.call(rbind, sigi_abf)
sigi_abf[, State := plyr::mapvalues(State, c("Sington", "Mixtrue"), c("State 1", "State 2"))]
sigi_abf[, State := factor(State, levels = c("State 1", "State 2"))]
sigi_abf[, A := gsub("CbC", "CMC", A)]
sigi_abf[, AA := gsub("CbC", "CMC", AA)]
```

```{r}
sigi_abf <- split(sigi_abf, sigi_abf$ID)
sigi_abf <- lapply(sigi_abf, function(x) {
  x$Time <- x$Time - min(x$Time)
  x$pA2 <- x$pA / sigi[ID %in% x$ID, BaseMean]
  x
})
sigi_abf <- do.call(rbind, sigi_abf)
```

```{r}
sigi_abf12 <- sigi_abf[A %in% c("Cys", "CMC")]
sigi_abf12[, A := factor(A, levels = c("Cys", "CMC"))]
```

```{r}
ggplot(sigi_abf12[State == "State 1"], aes(x = Time * 1000, y = pA2, colour = A)) + 
  geom_step() + 
  facet_grid(. ~ A, scales = "free", space = "free") + 
  scale_y_continuous(n.breaks = 4) + 
  scale_x_continuous(breaks = c(0, 1)) + 
  labs(x = "Time (ms)", y = "Normalized current") + 
  theme_light(base_size = 15) + 
  theme(panel.grid = element_blank(), 
        plot.background = element_blank(), 
        strip.background = element_blank(), 
        strip.text = element_text(colour = "black"), 
        legend.position = "none") -> p3
p3
```


```{r}
ggplot(sigi_abf12[State == "State 2"], aes(x = Time * 1000, y = pA2, colour = A)) + 
  geom_step() + 
  facet_grid(. ~ A, scales = "free", space = "free") + 
  scale_y_continuous(n.breaks = 4) + 
  scale_x_continuous(breaks = c(0, 1)) + 
  labs(x = "Time (ms)", y = "Normalized current") + 
  theme_light(base_size = 15) + 
  theme(panel.grid = element_blank(), 
        plot.background = element_blank(), 
        strip.background = element_blank(), 
        strip.text = element_text(colour = "black"), 
        legend.position = "none") -> p4
p4
```

```{r fig.width=10, fig.height=6}
cowplot::plot_grid(p1, p3, p2, p4)
p1 + p3 + p2 + p4
```
