==> 90.UnnaturalPTM/13.Ack_Signal_Select.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(dplyr)
library(lsa)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    if(is(dists, "kNN")) {
      if(k > dists$k) stop("K too big")
      knn1 <- dists
      knn2 <- apply(knn1$id[, 1:k], 2, function(x) x %in% which(grepl("^Noise", rownames(knn1$id))))
    } else {
      knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
      knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
    }
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
AABlockade <- lapply(list.files("./analysis/61.SignalSelecting/01.StandardAA/02.SelectedSignals", "_V2.Rds", full.names = T), function(x) readRDS(x)$Summary)
AABlockade <- do.call(rbind, AABlockade)
AABlockade <- AABlockade[State == "State1"]
```

# AcK

```{r}
meta <- as.data.table(openxlsx::read.xlsx("data/ChenShanchuan/20231027AcK/20231027_AcK2.xlsx", sheet = 1))[, 1:7]
colnames(meta) <- c("file_name", "start_time", "end_time",  "amino_acid", "file_id", "sample", "base_line")
meta <- meta[grepl("非天然氨基酸AcK", sample)]
meta[, file_name := gsub(".abf", "", file_name)]
setkey(meta, file_name, start_time)
meta <- meta[, .SD[, .(amino_acid = base_line, start_time, end_time, file_id = paste0(file_name, ".", seq_len(.N)))], file_name]
meta[, start_time := as.numeric(start_time)]
meta[, end_time := as.numeric(end_time)]
meta[, sig_file := paste0("./analysis/90.UnnaturalPTM/01.SelectedL0/", file_id, ".MainL0.txt")]
meta[, file.exists(sig_file)]
meta <- meta[file.exists(sig_file)]
```

```{r}
meta
```

```{r fig.width=24, fig.height=6}
dataset <- do.call(rbind, lapply(meta[amino_acid != "blank", sig_file], fread))
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", unique(meta[amino_acid != "blank", file_name]), ".txt")
FeatureMatrix <- do.call(rbind, lapply(B_file, fread))
FeatureMatrix <- FeatureMatrix[ID %in% dataset$ID]
FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | colnames(FeatureMatrix) %in% c("DwellTime", "DeltaMean", "StageSD", "CurrentSD", "SignalCurrentPercent", "SignalCurrentWidth", "Blockade"), with = F], row.names = FeatureMatrix[[1]])
```

# One by one

## 20231027_0017

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/90.UnnaturalPTM/02.SignalsDistance/20231027_0017.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231027_0017.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("K", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.9)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/90.UnnaturalPTM/03.SelectedSignals/20231027_0017.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231027_0018

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/90.UnnaturalPTM/02.SignalsDistance/20231027_0018.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231027_0018.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("K", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.9)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/90.UnnaturalPTM/03.SelectedSignals/20231027_0018.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231027_0025

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/90.UnnaturalPTM/02.SignalsDistance/20231027_0025.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231027_0025.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("K", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.9)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/90.UnnaturalPTM/03.SelectedSignals/20231027_0025.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231027_0026

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/90.UnnaturalPTM/02.SignalsDistance/20231027_0026.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231027_0026.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("K", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.88)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/90.UnnaturalPTM/03.SelectedSignals/20231027_0026.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```











# All

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/90.UnnaturalPTM/02.SignalsDistance/AcK_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset, dists = alldists, k = 15, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("K", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  # scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset, aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/90.UnnaturalPTM/03.SelectedSignals/AcK_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```
